chore: 移动路由相对路径转换 - 处理 VueCropper 打开报错问题
