+++ src/views/account/settings/AvatarModal.vue
@@ -55,12 +55,7 @@
 
 </template>
 <script>
-import { VueCropper } from 'vue-cropper'
-
 export default {
-  components: {
-    VueCropper
-  },
   data () {
     return {
       visible: false,
